travis-worker initContainers fail with network problem
The initContainer is vault-sidekick, which tries to write
the vault info as secrets to our own kube controller.
Because we exclude 172.16.0.0/12, it fails and the
entire pod won't start up.

This will create the vault-sidekick as a job, and change
the travis-worker initContainer to check for the secrets mount.

Issue: org-ids/roadmap#7456